% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{apply_named_args}
\alias{apply_named_args}
\title{a version of apply() that matches the dimnames of X to the arguments of FUN}
\usage{
apply_named_args(X, MARGIN, FUN)
}
\arguments{
\item{X}{see documentation for apply}

\item{MARGIN}{see documentation for apply}

\item{FUN}{see documentation for apply}
}
\value{
see documentation for apply
}
\description{
a version of apply() that matches the dimnames of X to the arguments of FUN
}
