% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{enumerate_popn}
\alias{enumerate_popn}
\title{take a vector with the number of virions of each strain, and return a vector
of length equal to the number of virions, where each entry is the strain number
of that virion (the reverse of sum_strains)}
\usage{
enumerate_popn(virus_popn)
}
\arguments{
\item{virus_popn}{numeric vector of length 4, with the number of virions
in each of the 4 strains}
}
\value{
numeric vector of length sum(virus_popn), where each entry is the 
strain number of a virion
}
\description{
take a vector with the number of virions of each strain, and return a vector
of length equal to the number of virions, where each entry is the strain number
of that virion (the reverse of sum_strains)
}
