% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get_hash}
\alias{get_hash}
\title{get short hash of current commit}
\usage{
get_hash(repo_dir)
}
\arguments{
\item{repo_dir}{repo directory}
}
\value{
character vector containing short hash of current commit
}
\description{
get short hash of current commit
}
