% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{lapply_w_name}
\alias{lapply_w_name}
\title{a version of lapply() that supplies FUN with both the name and the value of each component}
\usage{
lapply_w_name(X, FUN)
}
\arguments{
\item{X}{see documentation for lapply}

\item{FUN}{function which takes a character vector of length 1 as a first argument and 
something else as a second argument}
}
\value{
see documentation for lapply
}
\description{
a version of lapply() that supplies FUN with both the name and the value of each component
}
