% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{make_and_package_segments}
\alias{make_and_package_segments}
\title{returns the number of newly produced virions after reassortment and before mutation}
\usage{
make_and_package_segments(prob_strain, burst_size)
}
\arguments{
\item{prob_strain}{numeric vector of length 4.  Probabilities of the segments
corresponding to each of the four strains being produced.}

\item{burst_size}{burst size from a given cell}
}
\value{
numeric vector of length 4, with the number of newly produced virions
in each of the 4 strains
}
\description{
returns the number of newly produced virions after reassortment and before mutation
}
