% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{make_mutation_matrix}
\alias{make_mutation_matrix}
\title{make matrix of probabilities of strain x mutating into strain y}
\usage{
make_mutation_matrix(mutation_prob)
}
\arguments{
\item{mutation_prob}{mutation probability}
}
\value{
4x4 matrix of mutation probabilities
}
\description{
make matrix of probabilities of strain x mutating into strain y
}
