% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{mutate_popn}
\alias{mutate_popn}
\title{mutate a population of virions}
\usage{
mutate_popn(virus_popn, mutation_prob)
}
\arguments{
\item{virus_popn}{numeric vector of length 4, with the number of virions
in each of the 4 strains}

\item{mutation_prob}{probability of mutation for each virion.  Assume same
for pb1 and pa.}
}
\value{
numeric vector of length 4, with the number of virions
in each of the 4 strains, after mutation
}
\description{
mutate a population of virions
}
