% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{mutate_popn_wrapper}
\alias{mutate_popn_wrapper}
\title{make function to mutate a population of virions}
\usage{
mutate_popn_wrapper(mutation_prob)
}
\arguments{
\item{mutation_prob}{probability of mutation for each virion.  Assume same
for pb1 and pa.}
}
\value{
function that takes the argument virus_popn: numeric vector of 
length 4, with the number of virions in each of the 4 strains, and outputs
the same vector but with mutated virions
}
\description{
make function to mutate a population of virions
}
