% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{plot_segments}
\alias{plot_segments}
\title{plot results of run_reassortment_model by segment}
\usage{
plot_segments(results)
}
\arguments{
\item{results}{output of run_reassortment_model}
}
\description{
plot results of run_reassortment_model by segment
}
