% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_strains}
\alias{plot_strains}
\title{plot results of run_xxx_model by strain}
\usage{
plot_strains(results)
}
\arguments{
\item{results}{output of run_xxx_model}
}
\description{
plot results of run_xxx_model by strain
}
