% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{probabilistic_round}
\alias{probabilistic_round}
\title{round a (vector of) numbers up with probability equal to their non-integer parts}
\usage{
probabilistic_round(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector of the same length as x
}
\description{
round a (vector of) numbers up with probability equal to their non-integer parts
}
