% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{reassort_popn}
\alias{reassort_popn}
\title{old implementation of reassortment}
\usage{
reassort_popn(virus_popn)
}
\arguments{
\item{virus_popn}{numeric vector of length 4, with the number of virions
in each of the 4 strains}
}
\value{
numeric vector of length 4, with the number of virions
in each of the 4 strains, after reassortment
}
\description{
old implementation of reassortment
}
