% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{run_default_pars}
\alias{run_default_pars}
\title{Simulate evolution of mixture by reassortment and/or mutation, for many replicates}
\usage{
run_default_pars(MOI = 1, iv = c(95, 0, 0, 5), fitness_MW = 0,
  fitness_WM = 1.25, fitness_MM = 1, mutation_prob = 2e-04,
  reassort = TRUE, pop_size = 1e+06, burst_size = 10,
  MOI_dependent_burst_size = TRUE, choose_strain_by_fitness = FALSE,
  sim_name, hash)
}
\arguments{
\item{MOI}{multiplicity of infection = number of virions/number of cells}

\item{iv}{numeric vector of length n_strains = 4.
Initial proportions of [mt, mt], [mt, wt], [wt, mt], [wt, wt].}

\item{fitness_MW}{fitness of the PB1 mutant}

\item{fitness_WM}{fitness of the PA mutant}

\item{fitness_MM}{fitness of the double mutant}

\item{mutation_prob}{probability of mutation for each new virion.  Assume same
for pb1 and pa.}

\item{reassort}{logical. If TRUE, random reassortment of segments occurs
before packaging.  If FALSE, each strain in the cell effectively produces
pre-packaged viruses}

\item{pop_size}{numeric vector of length 1. Number of virions.}

\item{burst_size}{numeric vector of length 1.
Burst size from a cell infected with one virion with fitness 1.}

\item{sim_name}{character vector for naming of output file}

\item{hash}{character vector: git hash for tagging run}
}
\value{
matrix of dim c(generations * n_replicates, n_strains + 1).  
Concatenated simulation results for n_replicates simulations with the
above model parameters.
}
\description{
Simulate evolution of mixture by reassortment and/or mutation, for many replicates
}
