% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{run_mutation_model_no_coinfection}
\alias{run_mutation_model_no_coinfection}
\title{Simulate evolution of mixture by reassortment}
\usage{
run_mutation_model_no_coinfection(iv, fitness, burst_size, pop_size,
  generations, mutation_prob)
}
\arguments{
\item{iv}{numeric vector of length n_strains = 4.
Initial proportions of [mt, mt], [mt, wt], [wt, mt], [wt, wt].}

\item{fitness}{numeric vector of length n_strains = 4.
Relative fitness of the four strains.}

\item{burst_size}{numeric vector of length 1.
Burst size from a cell infected with one virion with fitness 1.}

\item{pop_size}{numeric vector of length 1. Initial number of virions.}

\item{generations}{numeric vector of length 1. Number of generations for
which to run model.}

\item{mutation_prob}{probability of mutation for each new virion.  Assume same
for pb1 and pa.}
}
\value{
matrix of dim c(generations, n_strains).  Proportion of virions of
each strain for each generation.
}
\description{
Simulate evolution of mixture by reassortment
}
