% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{run_reassortment_model}
\alias{run_reassortment_model}
\title{Simulate evolution of mixture by reassortment}
\usage{
run_reassortment_model(iv, fitness, burst_size, pop_size, max_pop_size,
  generations)
}
\arguments{
\item{iv}{numeric vector of length n_strains = 4.
Initial proportions of [mt, mt], [mt, wt], [wt, mt], [wt, wt].}

\item{fitness}{numeric vector of length n_strains = 4.
Relative fitness of the four strains.}

\item{burst_size}{numeric vector of length 1.
Burst size from a cell infected with one virion with fitness 1.}

\item{pop_size}{numeric vector of length 1. Initial number of virions.}

\item{max_pop_size}{numeric vector of length 1. maximum number of virions.}

\item{generations}{numeric vector of length 1. Number of generations for
which to run model.}
}
\value{
matrix of dim c(generations, n_strains).  Proportion of virions of
each strain for each generation.
}
\description{
Simulate evolution of mixture by reassortment
}
