% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{segments_to_strain}
\alias{segments_to_strain}
\title{turns a list of segment indices into a strain number}
\usage{
segments_to_strain(idx)
}
\arguments{
\item{idx}{numeric vector of length 2, where each entry is 1 or 0: 
1 in the first entry indicates MT PB1,
1 in the second entry indicates MT PA}
}
\value{
the strain number of the virion -- MM = 1, MW = 2, WM = 3, WW = 4
}
\description{
turns a list of segment indices into a strain number
}
