% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sensitivity_plot}
\alias{sensitivity_plot}
\title{plot proportion of each strain at 20 generations across sensitivity analyses}
\usage{
sensitivity_plot(results_list, results_name, reassortment_in, x_lab,
  rotate_x_labels = FALSE)
}
\arguments{
\item{results_list}{list of results; each element is an output of run_xxx_model}

\item{results_name}{vector of names of simulation to plot along x-axis}

\item{reassortment_in}{vector of logicals: whether those results were from a simulation
with or without reassortment}
}
\description{
plot proportion of each strain at 20 generations across sensitivity analyses
}
