% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tools.R
\name{setup_cluster}
\alias{setup_cluster}
\title{Cluster setup}
\usage{
setup_cluster(n_cores = 1)
}
\arguments{
\item{n_cores}{number of cluster cores}
}
\value{
a didehpc::queue_didehpc object
}
\description{
This function sets up an object linking to the DIDE cluster in a very crude way.
}
