% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{simulate_evolution}
\alias{simulate_evolution}
\title{Simulate evolution of mixture by reassortment and/or mutation}
\usage{
simulate_evolution(iv, fitness, burst_size, n_cells, pop_size, generations,
  mutation_prob, coinfection, MOI_dependent_burst_size,
  choose_strain_by_fitness, one_strain_produced, reassort)
}
\arguments{
\item{iv}{numeric vector of length n_strains = 4.
Initial proportions of [mt, mt], [mt, wt], [wt, mt], [wt, wt].}

\item{fitness}{numeric vector of length n_strains = 4.
Relative fitness of the four strains.}

\item{burst_size}{numeric vector of length 1.
Burst size from a cell infected with one virion with fitness 1.}

\item{n_cells}{numeric vector of length 1.  Number of cells.}

\item{pop_size}{numeric vector of length 1. Number of virions.}

\item{generations}{numeric vector of length 1. Number of generations for
which to run model.}

\item{mutation_prob}{probability of mutation for each new virion.  Assume same
for pb1 and pa.}

\item{coinfection}{logical.  If TRUE, cells can be infected by more than one
virion.  If FALSE, each cell can only be infected by one virion.}

\item{MOI_dependent_burst_size}{logical.  If TRUE, for a cell infected with a
single strain, the burst size scales with the MOI for that cell. If FALSE,
the burst size is independent of the MOI.}

\item{choose_strain_by_fitness}{logical.  If TRUE, strain(s) producing virions
in cell is/are chosen proportional to fitness and abundance, otherwise 
proportional to abundance only}

\item{one_strain_produced}{logical. If TRUE, each cell can only produce progeny
from one of its co-infecting strains.  If FALSE, each cell can produce progeny
for all of its coinfecting strains.}

\item{reassort}{logical. If TRUE, random reassortment of segments occurs
before packaging.  If FALSE, each strain in the cell effectively produces
pre-packaged viruses}
}
\value{
matrix of dim c(generations, n_strains).  Proportion of virions of
each strain for each generation.
}
\description{
Simulate evolution of mixture by reassortment and/or mutation
}
