% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment_model.R
\name{sum_strains}
\alias{sum_strains}
\title{takes a vector of length equal to the number of virions, 
where each entry is the strain number of that virion, and returns a vector 
with the number of virions of each strain (the reverse of enumerate_popn)}
\usage{
sum_strains(strains)
}
\arguments{
\item{strains}{a vector of length equal to the number of virions, 
where each entry is the strain number of that virion}
}
\value{
numeric vector of length 4, with the number of virions
in each of the 4 strains
}
\description{
takes a vector of length equal to the number of virions, 
where each entry is the strain number of that virion, and returns a vector 
with the number of virions of each strain (the reverse of enumerate_popn)
}
