% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{vnapply}
\alias{vnapply}
\title{vapply returning numeric vector of length 1}
\usage{
vnapply(X, FUN, ...)
}
\arguments{
\item{X}{see vapply}

\item{FUN}{see vapply}

\item{...}{additional arguments to pass to vapply}
}
\value{
output of vapply
}
\description{
vapply returning numeric vector of length 1
}
