% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{make_results_folder}
\alias{make_results_folder}
\title{make folder to store results and (optional) store hash of repo}
\usage{
make_results_folder(folder_name, base_dir, hash)
}
\arguments{
\item{folder_name}{folder name}

\item{base_dir}{directory in which to make folder}

\item{hash}{string -- make file with this name in folder}
}
\value{
name of created directory
}
\description{
make folder to store results and (optional) store hash of repo
}
