% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_segments}
\alias{plot_segments}
\title{plot results of run_xxx_model by segment}
\usage{
plot_segments(results)
}
\arguments{
\item{results}{output of run_xxx_model}
}
\description{
plot results of run_xxx_model by segment
}
